// src/modules/influencer/application/dtos/request/influencer-response.dto.ts
export interface InfluencerResponse {
  id: string;
  name: string;
  ageRange: string;
  gender: string;
  bodyShape: string;
  skinColor: string;
  eyeColor: string;
  hairType: string;
  faceType: string;
  country: string;
  lipsType: string;
  hairColor: string;
  height: number;
  influencerUrlImage: string;
  createDate: Date; 
  sizeImage: string;
}